Add a comptime known size string to the String struct
This allows the user to allocate a predefined size. Usefull for strnigs
that dont grow and where the size can be esitmated.

As a consequecnce the interfaces change to take pointers to the String
instead of the String itselfs as taking the string by copy would be a
performence degradation when the predefined string is not a
insignificatnt size.

Also change cmp to return the std.math.order instead of a bool
New function: eql implements old cmp.

Fix usage